+++ playground/modules/excluded-module/excluded.stories.js
@@ -1,0 +1,5 @@
+export default {
+    title: "Excluded Story"
+}
+
+export const NO = () => "<div>You can't see me</div>"