+++ playground/modules/included-module/included.stories.js
@@ -1,0 +1,5 @@
+export default {
+    title: "Included Story"
+}
+
+export const YES = () => "<div>YES</div>"